LayoutTests/crypto/webkitSubtle/aes-cbc-encrypt-decrypt-with-padding-expected.txt

CONSOLE MESSAGE: line 22: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test AES-CBC encrypt and decrypt functions on a plaintext that is not a multiple of block size in length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ['decrypt', 'encrypt']
Using the key to encrypt plaintext...
PASS bytesToHexString(new Uint8Array(cyphertext)) is '630199c5f202cc7167bb84c6c72b349d'
Decrypting it back...
PASS new Uint8Array(decryptionResult) is plaintext
PASS successfullyParsed is true

TEST COMPLETE
